购物车选取规格问题

+++ ZFB/Class/Main/Home/SearchBarController/Model/SearchNoResultModel.h
@@ -76,7 +76,7 @@
 
 @property (nonatomic, copy) NSString *storeDist;
 
-@property (nonatomic, assign) NSInteger storePrice;
+@property (nonatomic, assign) CGFloat storePrice;
 
 @property (nonatomic, assign) NSInteger auditStatus;
